extract derive macro into a separated crate
