+++ derive_builder_macro/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "derive_builder"
+name = "derive_builder_macro"
 version = "0.10.0-alpha"
 authors = ["Colin Kiegel <kiegel@gmx.de>",
            "Pascal Hertleif <killercup@gmail.com>",
@@ -8,30 +8,30 @@ authors = ["Colin Kiegel <kiegel@gmx.de>",
 
 description = "Rust macro to automatically implement the builder pattern for arbitrary structs."
 repository = "https://github.com/colin-kiegel/rust-derive-builder"
-documentation = "https://docs.rs/derive_builder/0.10.0-alpha"
+documentation = "https://docs.rs/derive_builder_macro/0.10.0-alpha"
 
 license = "MIT/Apache-2.0"
 categories = ["development-tools", "rust-patterns"]
 keywords = ["derive", "macro", "builder", "setter", "struct"]
 readme = "README.md"
-build = "build/mod.rs"
 
 [badges]
 travis-ci = { repository = "colin-kiegel/rust-derive-builder" }
 
+[lib]
+proc-macro = true
+
 [features]
-logging = [ "derive_builder_macro/logging", "derive_builder_core/logging" ]
+logging = [ "log", "env_logger", "derive_builder_core/logging" ]
 skeptic_tests = ["skeptic"]
 clippy = ["derive_builder_core/clippy"]
 
 [dependencies]
-derive_builder_macro = { version = "=0.10.0-alpha", path = "../derive_builder_macro" }
+darling = "0.10.2"
+proc-macro2 = "1.0"
+syn = { version = "1", features = ["full", "extra-traits"] }
+quote = "1.0"
+log = { version = "0.4", optional = true }
+env_logger = { version = "0.5", optional = true }
 derive_builder_core = { version = "=0.10.0-alpha", path = "../derive_builder_core" }
 skeptic = { version = "0.13", optional = true }
-
-[build-dependencies]
-skeptic = { version = "0.13", optional = true }
-
-[dev-dependencies]
-pretty_assertions = "0.6"
-compiletest_rs = "0.5"
